# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hotel Booking Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: roomId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    delete:
      operationId: deleteBooking
      tags: [bookings]
      summary: Delete a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /bookings/{id}/cancel:
    patch:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
...
  /auth/login:
    post:
      operationId: login
      summary: Authenticate user
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentials"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      summary: Get current session user
      tags: [auth]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
  /auth/sessions:
    get:
      operationId: listActiveSessions
      summary: List active sessions (admin/supervisor only)
      tags: [auth]
      responses:
        "200":
          description: Active sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActiveSessionList"
        "401":
          description: Not authenticated
        "403":
          description: Forbidden
...
components:
  schemas:
...
    Booking:
      type: object
      required: [id, guestName, guestEmail, roomId, checkIn, checkOut, status, totalAmount, createdAt]
      properties:
        id:
          type: integer
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: ["string", "null"]
        roomId:
          type: integer
        roomName:
          type: ["string", "null"]
        roomType:
          type: ["string", "null"]
        checkIn:
          type: string
        checkOut:
          type: string
        status:
          type: string
          description: "confirmed, checked-in, checked-out, cancelled"
        totalAmount:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    BookingInput:
      type: object
      required: [guestName, guestEmail, roomId, checkIn, checkOut, totalAmount]
      properties:
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: string
        roomId:
          type: integer
        checkIn:
          type: string
        checkOut:
          type: string
        status:
          type: string
        totalAmount:
          type: number
        notes:
          type: string
    BookingUpdate:
      type: object
      properties:
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: string
        roomId:
          type: integer
        checkIn:
          type: string
        checkOut:
          type: string
        status:
          type: string
        totalAmount:
          type: number
        notes:
          type: string
...
    AuthCredentials:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthUser:
      type: object
      required: [id, username, displayName, role, permissions, isActive]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        email:
          type: string
          nullable: true
        role:
          type: string
        permissions:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        createdAt:
          type: string
    ActiveSession:
      type: object
      required: [sessionKey, userId, displayName, username, role]
      properties:
        sessionKey:
          type: string
        userId:
          type: integer
        displayName:
          type: string
        username:
          type: string
        role:
          type: string
    ActiveSessionList:
      type: object
      required: [sessions]
      properties:
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/ActiveSession"
...